+++ lib/src/headless_in_app_webview.dart
@@ -197,7 +197,7 @@ class HeadlessInAppWebView implements WebView {
       onConsoleMessage;
 
   @override
-  final Future<WebView> Function(InAppWebViewController controller,
+  final Future<bool> Function(InAppWebViewController controller,
       CreateWindowRequest onCreateWindowRequest) onCreateWindow;
 
   @override